+++ lib/elasticity.rb
@@ -6,8 +6,8 @@ require 'nokogiri'
 
 require 'elasticity/aws_request'
 require 'elasticity/emr'
-require 'elasticity/job_flow'
-require 'elasticity/job_flow_step'
+require 'elasticity/job_flow_status'
+require 'elasticity/job_flow_status_step'
 
 require 'elasticity/simple_job'