+++ spec/lib/elasticity/job_flow_status_step_spec.rb
@@ -1,4 +1,4 @@
-describe Elasticity::JobFlowStep do
+describe Elasticity::JobFlowStatusStep do
 
   before do
     describe_jobflows_xml = <<-JOBFLOWS
@@ -51,7 +51,7 @@ describe Elasticity::JobFlowStep do
 
   describe ".from_xml" do
     it "should return a JobFlowStep with the appropriate fields initialized" do
-      jobflow_step = Elasticity::JobFlowStep.from_member_element(@members_nodeset[0])
+      jobflow_step = Elasticity::JobFlowStatusStep.from_member_element(@members_nodeset[0])
       jobflow_step.name.should == "Setup Hive"
       jobflow_step.state.should == "FAILED"
       jobflow_step.started_at.should == Time.parse("2011-10-04T21:49:16Z")
@@ -61,7 +61,7 @@ describe Elasticity::JobFlowStep do
 
   describe ".from_steps_nodeset" do
     it "should return JobFlowSteps with the appropriate fields initialized" do
-      jobflow_steps = Elasticity::JobFlowStep.from_members_nodeset(@members_nodeset)
+      jobflow_steps = Elasticity::JobFlowStatusStep.from_members_nodeset(@members_nodeset)
       jobflow_steps.map(&:name).should == ["Setup Hive", "Run Hive Script"]
       jobflow_steps.map(&:state).should == ["FAILED", "PENDING"]
       jobflow_steps.map(&:started_at).should == [Time.parse("2011-10-04T21:49:16Z"), nil]